Feature: move ecies interface code into go-ecies
The PKCS11 related code remains in this package.
