+++ internal/ecies.go
@@ -8,17 +8,17 @@ import (
 	"fmt"
 
 	"github.com/ThalesIgnite/crypto11"
-	"github.com/umbracle/ecies"
+	ecies "github.com/foundriesio/go-ecies"
 )
 
 type EciesCrypto struct {
-	PrivKey PrivateKey
+	PrivKey ecies.KeyProvider
 	ctx     *crypto11.Context
 }
 
 func NewEciesLocalHandler(privKey crypto.PrivateKey) CryptoHandler {
 	if ec, ok := privKey.(*ecdsa.PrivateKey); ok {
-		return &EciesCrypto{ImportECDSA(ec), nil}
+		return &EciesCrypto{ecies.ImportECDSA(ec), nil}
 	}
 	return nil
 }
@@ -28,7 +28,7 @@ func (ec *EciesCrypto) Decrypt(value string) ([]byte, error) {
 	if err != nil {
 		return nil, fmt.Errorf("Unable to base64 decode: %v", err)
 	}
-	decrypted, err := EciesDecrypt(ec.PrivKey, data, nil, nil)
+	decrypted, err := ecies.Decrypt(ec.PrivKey, data, nil, nil)
 	if err != nil {
 		return nil, fmt.Errorf("Unable to ECIES decrypt %v", err)
 	}
@@ -51,4 +51,24 @@ func (ec *EciesCrypto) Close() {
 
 func NewEciesPkcs11Handler(ctx *crypto11.Context, privKey crypto11.Signer) CryptoHandler {
 	return &EciesCrypto{ImportPcks11(ctx, privKey), ctx}
+}
+
+type PrivateKeyPkcs11 struct {
+	*ecies.PublicKey
+	ctx    *crypto11.Context
+	signer crypto11.Signer
+}
+
+func ImportPcks11(ctx *crypto11.Context, privKey crypto.PrivateKey) *PrivateKeyPkcs11 {
+	signer := privKey.(crypto11.Signer)
+	pub := signer.Public().(*ecdsa.PublicKey)
+	return &PrivateKeyPkcs11{ecies.ImportECDSAPublic(pub), ctx, signer}
+}
+
+func (prv *PrivateKeyPkcs11) GenerateShared(pub *ecies.PublicKey) (sk []byte, err error) {
+	return prv.ctx.ECDH1Derive(prv.signer, pub.ExportECDSA())
+}
+
+func (prv *PrivateKeyPkcs11) Public() *ecies.PublicKey {
+	return prv.PublicKey
 }
